requirements.txt
aiohttp==3.8.4
#azure_storage==0.37.0
channels==4.0.0
# chromadb==0.3.22
# Django==4.1.5
# docx==0.2.4
#faiss==1.5.3
faiss_cpu==1.7.4
fire==0.4.0
# godot==0.1.1
# google_api_python_client==2.93.0
lancedb==0.1.16
langchain~=0.0.343
loguru==0.6.0
meilisearch==0.21.0
numpy==1.24.3
openai==0.27.8
openpyxl
beautifulsoup4==4.12.2
pandas==2.0.3
pydantic~=1.10.8
#pygame==2.1.3
#pymilvus==2.2.8
pytest==7.2.2
python_docx==0.8.11
PyYAML==6.0
# sentence_transformers==2.2.2
setuptools==65.6.3
tenacity==8.2.2
tiktoken==0.4.0
tqdm==4.64.0
#unstructured[local-inference]
# playwright
# selenium>4
# webdriver_manager<3.9
anthropic==0.3.6
typing-inspect==0.8.0
typing_extensions==4.5.0
libcst==1.0.1
qdrant-client==1.4.0
pytest-mock==3.11.1
open-interpreter==0.1.4; python_version>"3.9"
ta==0.10.2
semantic-kernel==0.3.10.dev0
websocket-client==0.58.0
amazoncaptcha
aiofiles~=23.2.1
pygments~=2.16.1
requests~=2.31.0
yaml~=0.2.5
streamlit
google-cloud-aiplatform
pandas
google_cloud_discoveryengine
fastapi
uvicorn==0.24.0.post1
firebase-admin
sh~=2.0.6
sse-starlette==1.8.2
starlette==0.33.0
h11==0.14.0